authentication and authorization

+++ Controllers/UsersController.cs
@@ -4,9 +4,15 @@ using Social_Network_API.Entities;
 using Microsoft.AspNetCore.Http;
 using Social_Network_API.Models;
 using Social_Network_API.Database;
+using Microsoft.AspNetCore.Authorization;
+using System.Security.Claims;
+using Social_Network_API.Enums;
 
 namespace Social_Network_API.Controllers
 {
+
+
+    
     
     [Route("user")]
     [ApiController]
@@ -22,7 +28,47 @@ namespace Social_Network_API.Controllers
             
         }
 
-        
+        private User? GetCurrentUser()
+        {
+
+            ClaimsIdentity? identity = HttpContext.User.Identity as ClaimsIdentity;
+            
+            if (identity != null)
+            {
+
+                var userClaims = identity.Claims;
+                int idToSearch = Convert.ToInt32(userClaims.FirstOrDefault(x => x.Type == ClaimTypes.NameIdentifier)?.Value);
+                
+
+                if(!_context.Users.Any(x => x.Id == idToSearch))
+                {
+                    return null;
+                }
+                User? temp = _context.Users.Where(x => x.Id == idToSearch).ToArray().First();
+                if(temp != null)
+                {
+                    return temp;
+                }
+                
+            }
+            return null;
+        }
+        [Route("me")]
+        [Authorize]
+        public IActionResult GetMe()
+        {
+
+            User? response = GetCurrentUser();
+            if(response!= null)
+            {
+                return new JsonResult(response);
+            }
+            return new StatusCodeResult(500);
+            
+        }
+
+
+        [AllowAnonymous]
         [Route("{id?}")]
         [HttpGet]
         public IActionResult Get([FromRoute] int id)
@@ -36,6 +82,7 @@ namespace Social_Network_API.Controllers
             
             return new JsonResult(tempUser.ToArray().First());
         }
+        [AllowAnonymous]
         [Route("count")]
         [HttpGet]
         public IActionResult Count()
@@ -44,9 +91,13 @@ namespace Social_Network_API.Controllers
         }
 
         [HttpGet]
-
-        public IActionResult Get([FromQuery] int index, int count)
+        [AllowAnonymous]
+        public IActionResult GetList([FromQuery] int index, int count)
         {
+            if(index <0 || count < 0)
+            {
+                return BadRequest();
+            }
             if (index == 0 && count == 0)
             {
                 return BadRequest();
@@ -67,16 +118,20 @@ namespace Social_Network_API.Controllers
 
         }
 
-
+        [Authorize]
+        
         [Route("{id?}")]
         [HttpDelete]
         public IActionResult Delete([FromRoute] int id)
         {
-
+            if(GetCurrentUser().Role != UserRole.Admin)
+            {
+                return StatusCode(403);
+            }
             if (!_context.Users.Any(e => e.Id == id))
             {
                 HttpContext.Response.StatusCode = 409;
-                return new JsonResult(new { description = $"no user with id{id}"});
+                return new JsonResult(new { description = $"no user with id {id}"});
             }
             _context.Users.Remove(_context.Users.Where(e=>e.Id == id).ToArray().First());
             _context.SaveChanges();
